requirements.txt
pandas
numpy
scikit-learn
catboost
matplotlib
seaborn
xgboost
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pyodbc>=5.0.0
python-dotenv>=1.0.0